Move to VS2013
git-svn-id: svn://localhost@998 5519e62b-7011-5a40-965f-37cedbea8841

+++ NEWin32/Win32.cpp
@@ -36,7 +36,7 @@ void wmain(int argc, wchar_t *argv[])
 
 			auto handleInfo = Win32Lib::GetHandleInfo(handles);
 			for each (auto handle in *handleInfo)
-				printf("PID: %i, Handle: %i, Type: %S, Name: %S, Data %S\n", handle->PID, handle->Handle, handle->Type.c_str(), handle->Name.c_str(), handle->Data.c_str());
+				printf("PID: %i, Handle: %i, Type: %S, Name: %S, Data %S\n", handle->PID, (intptr_t)handle->Handle, handle->Type.c_str(), handle->Name.c_str(), handle->Data.c_str());
 		}
 		else if (params[0] == L"-types")
 		{
